+++ muonShieldOptimization/exitHadronAbsorber.cxx
@@ -39,6 +39,8 @@ Double_t mm  = 0.1*cm;  //  mm
 
 exitHadronAbsorber::exitHadronAbsorber()
   : FairDetector("exitHadronAbsorber", kTRUE, kVETO),
+    fUniqueID(-1),
+    fEventId(-1),
     fTrackID(-1),
     fVolumeID(-1),
     fPos(),
@@ -68,7 +70,9 @@ Bool_t  exitHadronAbsorber::ProcessHits(FairVolume* vol)
   /** This method is called from the MC stepping */
   if ( gMC->IsTrackEntering() ) {
     fTrackID  = gMC->GetStack()->GetCurrentTrackNumber();
+    fEventId = gMC->CurrentEvent();
     TParticle* p  = gMC->GetStack()->GetCurrentTrack();
+    fUniqueID = p->GetUniqueID();
     Int_t pdgCode = p->GetPdgCode();
     gMC->TrackMomentum(fMom);
     if (!(fOnlyMuons && TMath::Abs(pdgCode)!=13)){
@@ -78,7 +82,7 @@ Bool_t  exitHadronAbsorber::ProcessHits(FairVolume* vol)
      if ( (fMom.E()-fMom.M() )>EMax) {
       AddHit(fTrackID, 111, TVector3(fPos.X(),fPos.Y(),fPos.Z()),
            TVector3(fMom.Px(), fMom.Py(), fMom.Pz()), fTime, fLength,
-           0,pdgCode,TVector3(p->Vx(), p->Vy(), p->Vz()),TVector3(p->Px(), p->Py(), p->Pz()) );
+           0,pdgCode,TVector3(p->Vx(), p->Vy(), p->Vz()),TVector3(p->Px(), p->Py(), p->Pz()), fEventId);
       ShipStack* stack = dynamic_cast<ShipStack*>(gMC->GetStack());
       stack->AddPoint(kVETO);
       }
@@ -282,12 +286,12 @@ void exitHadronAbsorber::ConstructGeometry()
 vetoPoint* exitHadronAbsorber::AddHit(Int_t trackID, Int_t detID,
                                       TVector3 pos, TVector3 mom,
                                       Double_t time, Double_t length,
-                                      Double_t eLoss, Int_t pdgCode,TVector3 Lpos, TVector3 Lmom)
+                                      Double_t eLoss, Int_t pdgCode,TVector3 Lpos, TVector3 Lmom, Int_t eventID)
 {
   TClonesArray& clref = *fexitHadronAbsorberPointCollection;
   Int_t size = clref.GetEntriesFast();
   return new(clref[size]) vetoPoint(trackID, detID, pos, mom,
-         time, length, eLoss, pdgCode,Lpos,Lmom);
+         time, length, eLoss, pdgCode,Lpos,Lmom, eventID);
 }
 
 void exitHadronAbsorber::Register()