+++ veto/vetoPoint.cxx
@@ -20,8 +20,9 @@ vetoPoint::vetoPoint(Int_t trackID,
                      Double_t eLoss,
                      Int_t pdgcode,
                      TVector3 Lpos,
-                     TVector3 Lmom)
-    : FairMCPoint(trackID, detID, pos, mom, tof, length, eLoss)
+                     TVector3 Lmom,
+                     Int_t eventID)
+    : FairMCPoint(trackID, detID, pos, mom, tof, length, eLoss, eventID)
     , fPdgCode(pdgcode)
     , fLpos{Lpos.X(), Lpos.Y(), Lpos.Z()}
     , fLmom{Lmom.X(), Lmom.Y(), Lmom.Z()}