+++ veto/vetoPoint.h
@@ -35,7 +35,8 @@ class vetoPoint : public FairMCPoint
               Double_t eLoss,
               Int_t pdgCode,
               TVector3 Lpos,
-              TVector3 Lmom);
+              TVector3 Lmom,
+              Int_t eventID = -1);
 
     /** Destructor **/
     virtual ~vetoPoint();